undid some of the type moves into common that werent needed, cleaning up PR a bit
