+++ x-pack/plugins/security_solution/common/detection_engine/types.ts
@@ -53,57 +53,3 @@ export interface EqlSearchResponse<T> {
     events?: Array<BaseHit<T>>;
   };
 }
-
-export interface ShardsResponse {
-  total: number;
-  successful: number;
-  failed: number;
-  skipped: number;
-  failures?: ShardError[];
-}
-
-export interface ShardError {
-  shard: number;
-  index: string;
-  node: string;
-  reason: {
-    type: string;
-    reason: string;
-    index_uuid: string;
-    index: string;
-    caused_by: {
-      type: string;
-      reason: string;
-    };
-  };
-}
-
-export interface SearchResponse<T> {
-  took: number;
-  timed_out: boolean;
-  _scroll_id?: string;
-  _shards: ShardsResponse;
-  hits: {
-    total: TotalValue | number;
-    max_score: number;
-    hits: Array<
-      BaseHit<T> & {
-        _type: string;
-        _score: number;
-        _version?: number;
-        _explanation?: Explanation;
-        fields?: string[];
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        highlight?: any;
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        inner_hits?: any;
-        matched_queries?: string[];
-        sort?: string[];
-      }
-    >;
-  };
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  aggregations?: any;
-}
-
-export type SearchHit = SearchResponse<object>['hits']['hits'][0];
