+++ x-pack/plugins/security_solution/server/lib/detection_engine/signals/types.ts
@@ -14,8 +14,8 @@ import {
   AlertExecutorOptions,
   AlertServices,
 } from '../../../../../alerts/server';
+import { SearchResponse } from '../../types';
 import {
-  SearchResponse,
   EqlSearchResponse,
   BaseHit,
   RuleAlertAction,
