+++ x-pack/plugins/security_solution/server/lib/detection_engine/signals/utils.test.ts
@@ -46,7 +46,7 @@ import {
   sampleDocSearchResultsNoSortIdNoHits,
   repeatedSearchResultsWithSortId,
 } from './__mocks__/es_results';
-import { ShardError } from '../../../../common/detection_engine/types';
+import { ShardError } from '../../types';
 
 const buildRuleMessage = buildRuleMessageFactory({
   id: 'fake id',
