+++ x-pack/plugins/security_solution/server/lib/hosts/types.ts
@@ -17,8 +17,8 @@ import {
   DocValueFieldsInput,
 } from '../../graphql/types';
 import { FrameworkRequest, RequestOptionsPaginated } from '../framework';
-import { Hit, Hits } from '../types';
-import { SearchHit, TotalValue } from '../../../common/detection_engine/types';
+import { Hit, Hits, SearchHit } from '../types';
+import { TotalValue } from '../../../common/detection_engine/types';
 
 export interface HostsAdapter {
   getHosts(req: FrameworkRequest, options: HostsRequestOptions): Promise<HostsData>;
